Implements throttling functionality
This feature uses a hardcoded "blacklist" of `PROJECT_ID`s to selectively
throttle what events get sent back to the ProdPerfect servers.

IF (blacklist.contains(PROJECT_ID) {
  IF(should_this_session_uuid_be_sampled?()) {
     return continue_as_normal();
  } ELSE {
     return short_circuit_setup_logic_and_do_nothing_further();
  }
}

To determine whether or not a specific session should be sampled:

Calculate a MD5 hash from the `session_uuid`, and use the last 8 hex valued
characters of the hash to get a number (e.g. 0x499602d2 becomes the int 1234567890)
Compare this number to the highest possible 8 character hex value (0xffffffff or 4294967295) as a ratio:

1234567890/4294967295 = 0.2874.. or approx 29%

If we set our throttle value for this project ID to 50%, then the 29% outcome value would mean that we
*should* sample this session.

If we set our throttle value to something lower like 5%, then the 29% value would mean we would *not*
sample this value.

Since the `session_uuid` is a UUID and we use hashing, we can rely on the numerical values we extract
to be more-or-less normally distributed, and can use it as a means to decide whether or not the session
should be sampled.